% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRok.R
\name{grok_complete}
\alias{grok_complete}
\title{Chat Completions}
\usage{
grok_complete(messages)
}
\arguments{
\item{messages}{a list containing a users messages}
}
\value{
A list containing the API response from xAI
}
\description{
This function takes in a list object that contains a users message content,
model, stream, and temperature settings. To use the package you must have an
xAI API key stored as
Below is an example request.
}
\examples{
# example code
 message <-
c(list(messages=list(
list(role='system',content='You are a test assistant.'),
list(role='user',content='Testing. Just say hi and hello world and nothing else.')
)),
model='grok-beta',
stream=FALSE,
temperature=0)
grok_complete(message)
}
